src/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
gunicorn==20.1.0
